requirements.txt
pyyaml
python-dotenv
azure-cognitiveservices-speech
requests
sounddevice
numpy